Refactor Codeforces Polygon importer into utility class
